issues with bookings

+++ gateway/src/main/java/ru/practicum/shareit/client/BaseClient.java
@@ -55,8 +55,8 @@ public class BaseClient {
         return patch(path, null, null, body);
     }
 
-    protected <T> ResponseEntity<Object> patch(String path, long userId) {
-        return patch(path, userId, null, null);
+    protected <T> ResponseEntity<Object> patch(String path, Integer userId) {
+        return patch(path, Long.valueOf(userId), null, null);
     }
 
     protected <T> ResponseEntity<Object> patch(String path, long userId, T body) {
